Changing layout and code organization.
