+++ internal/pkg/cartoon/cartoons.go
